Returned more errors and created Azure Storage file
