+++ pkg/scalers/azure_eventhub.go
@@ -4,21 +4,18 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/url"
-	"strings"
 
 	eventhub "github.com/Azure/azure-event-hubs-go"
 	storageLeaser "github.com/Azure/azure-event-hubs-go/storage"
 	"github.com/Azure/azure-storage-blob-go/azblob"
 	"github.com/Azure/go-autorest/autorest/azure"
-	log "github.com/Sirupsen/logrus"
 )
 
 // GetStorageCredentials returns azure env and storage credentials
 func GetStorageCredentials(storageConnection string) (azure.Environment, *azblob.SharedKeyCredential, error) {
-	storageAccountName, storageAccountKey, err := parseAzureStorageConnectionString(storageConnection)
+	storageAccountName, storageAccountKey, err := ParseAzureStorageConnectionString(storageConnection)
 	if err != nil {
 		return azure.Environment{}, &azblob.SharedKeyCredential{}, fmt.Errorf("unable to parse connection string: %s\n", storageConnection)
 	}
@@ -72,7 +69,7 @@ func GetEventHubClient(connectionString string) (*eventhub.Hub, error) {
 
 // GetLeaseFromBlobStorage accesses Blob storage and gets lease information of a partition
 func GetLeaseFromBlobStorage(ctx context.Context, partitionID string, storageConnection string, storageContainerName string) (Lease, error) {
-	storageAccountName, _, err := parseAzureStorageConnectionString(storageConnection)
+	storageAccountName, _, err := ParseAzureStorageConnectionString(storageConnection)
 	if err != nil {
 		return Lease{}, fmt.Errorf("unable to parse storage connection string: %s", err)
 	}
@@ -89,8 +86,9 @@ func GetLeaseFromBlobStorage(ctx context.Context, partitionID string, storageCon
 
 	get, err := blobURL.Download(ctx, 0, 0, azblob.BlobAccessConditions{}, false)
 	if err != nil {
-		log.Fatal(err)
+		return Lease{}, fmt.Errorf("unable to download file from blob storage: %s", err)
 	}
+
 	blobData := &bytes.Buffer{}
 	reader := get.Body(azblob.RetryReaderOptions{})
 	blobData.ReadFrom(reader)
@@ -99,33 +97,8 @@ func GetLeaseFromBlobStorage(ctx context.Context, partitionID string, storageCon
 	var dat Lease
 
 	if err := json.Unmarshal(blobData.Bytes(), &dat); err != nil {
-		log.Fatalf("failed to decode blob data: %s\n", err)
+		return Lease{}, fmt.Errorf("failed to decode blob data: %s", err)
 	}
 
 	return dat, nil
-}
-
-// ParseAzureStorageConnectionString parses a storage account connection string into (accountName, key)
-func parseAzureStorageConnectionString(connectionString string) (string, string, error) {
-	parts := strings.Split(connectionString, ";")
-
-	var name, key string
-	for _, v := range parts {
-		if strings.HasPrefix(v, "AccountName") {
-			accountParts := strings.SplitN(v, "=", 2)
-			if len(accountParts) == 2 {
-				name = accountParts[1]
-			}
-		} else if strings.HasPrefix(v, "AccountKey") {
-			keyParts := strings.SplitN(v, "=", 2)
-			if len(keyParts) == 2 {
-				key = keyParts[1]
-			}
-		}
-	}
-	if name == "" || key == "" {
-		return "", "", errors.New("Can't parse connection string")
-	}
-
-	return name, key, nil
 }
